# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogicLens API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type SignupBodyRole =
  (typeof SignupBodyRole)[keyof typeof SignupBodyRole];

export const SignupBodyRole = {
  student: "student",
  admin: "admin",
} as const;

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  role?: SignupBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

/**
 * @nullable
 */
export type SavedTraceCustomInputs = { [key: string]: unknown } | null;

export interface SavedTrace {
  id: number;
  userId: number;
  title: string;
  category: string;
  subtype: string;
  code: string;
  /** @nullable */
  customInputs?: SavedTraceCustomInputs;
  /** @nullable */
  traceSummary?: string | null;
  /** @nullable */
  finalOutput?: string | null;
  savedAt: string;
}

/**
 * @nullable
 */
export type CreateTraceBodyCustomInputs = { [key: string]: unknown } | null;

export interface CreateTraceBody {
  title: string;
  category: string;
  subtype: string;
  code: string;
  /** @nullable */
  customInputs?: CreateTraceBodyCustomInputs;
  /** @nullable */
  traceSummary?: string | null;
  /** @nullable */
  finalOutput?: string | null;
}

export interface FavoriteProgram {
  id: number;
  userId: number;
  programId: number;
  programName: string;
  programCategory: string;
  addedAt: string;
}

export interface AddFavoriteBody {
  programId: number;
  programName: string;
  programCategory: string;
}

export interface SampleProgram {
  id: number;
  name: string;
  category: string;
  subtype: string;
  code: string;
  description: string;
  difficulty: string;
  featured: boolean;
  tags: string[];
}

export interface CreateProgramBody {
  name: string;
  category: string;
  subtype: string;
  code: string;
  description: string;
  difficulty: string;
  featured?: boolean;
  tags?: string[];
}

export interface CategorySummary {
  category: string;
  count: number;
  description: string;
}

export interface DashboardStats {
  totalTraces: number;
  totalFavorites: number;
  recentTraces: SavedTrace[];
  categoriesExplored: string[];
}

export interface AdminStats {
  totalUsers: number;
  totalTraces: number;
  totalFavorites: number;
  totalPrograms: number;
  programsByCategory: CategorySummary[];
}

export type GetProgramsParams = {
  category?: string;
  difficulty?: string;
};